requirements.txt
asgiref==3.5.2
Django==3.2.16
django-forms-bootstrap==3.1.0
djangorestframework==3.14.0
docutils==0.19
flake8==5.0.4
importlib-metadata==4.2.0
isort==5.11.5
mccabe==0.7.0
Pillow==9.3.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.6
sqlparse==0.4.3
typing-extensions==4.4.0
zipp==3.11.0
